Improved designs

+++ Core/views.py
@@ -34,7 +34,7 @@ class IndexView(View):
         i = 0
         members = Member.objects.filter(is_alumni=False).order_by('batch','name')
         if len(members) > 6:
-            members_lists = [members[i * 6: i * 6 + 6] for i in range(len(members) / 6)]
+            members_lists = [members[i * 6: i * 6 + 6] for i in range(int(len(members) / 6))]
             members_lists.append(members[i * 6 + 6:])
         else:
             members_lists = [members, ]
